+++ pkg/etop/api/admin/order.go
@@ -47,7 +47,7 @@ func GetOrders(ctx context.Context, q *wrapadmin.GetOrdersEndpoint) error {
 		return err
 	}
 	q.Result = &pborder.OrdersResponse{
-		Paging: pbcm.PbPageInfo(paging, query.Result.Total),
+		Paging: pbcm.PbPageInfo(paging, int32(query.Result.Total)),
 		Orders: pborder.PbOrdersWithFulfillments(query.Result.Orders, model.TagEtop, query.Result.Shops),
 	}
 	return nil
@@ -93,7 +93,7 @@ func GetFulfillments(ctx context.Context, q *wrapadmin.GetFulfillmentsEndpoint) 
 	}
 	q.Result = &pborder.FulfillmentsResponse{
 		Fulfillments: pborder.PbFulfillmentExtendeds(query.Result.Fulfillments, model.TagEtop),
-		Paging:       pbcm.PbPageInfo(paging, query.Result.Total),
+		Paging:       pbcm.PbPageInfo(paging, int32(query.Result.Total)),
 	}
 	return nil
 }
